Sprint 4 Changes
Added a Strafe mechanic - By holding left shift or left trigger, the player can walk at a reduced speed without turning around.

Added a healing spell. HP is accurately tracked internally, but currently faces issues reflecting this change on the heartbar aspect of the HUD

Attempted to add a spread-beam spell "Solar Ray". Currently does not spawn in the intended position, needs to be looked into.

Added a quickbind menu. By holding right-click or right trigger, the player can see a set of spell or item options that can be quickly accessed with a corresponding button press without the need to navigate menus. Currently missing indications of which button corresponds to which quickbind, as research needs to be done on how to detect which control scheme the player is using so that keys/buttons can be displayed accordingly